LayoutTests/http/tests/security/contentSecurityPolicy/media-src-allowed.html

<meta http-equiv="X-WebKit-CSP" content="media-src http://127.0.0.1:8000">
<video></video>
<script src=../../../media-resources/media-file.js></script>
<script src=../../../media-resources/video-test.js></script>
<script>
waitForEvent('loadedmetadata', function () {
    alert('PASS');
    endTestLater();
} );

// Find a supported media file.
var mediaFile = findMediaFile("video", "content/test");
var mimeType = mimeTypeForFile(mediaFile);

video.src = "http://127.0.0.1:8000/media/video-load-and-stall.cgi?name=../../../media/" + mediaFile + "&mimeType=" + mimeType + "&stallAt=100000";
</script>
