safari-534.53-branch/LayoutTests/http/tests/security/contentSecurityPolicy/media-src-allowed-expected.txt

ALERT: PASS
EVENT(loadedmetadata)
END OF TEST
